[refactor]別コンポーネントにした
